pyproject.toml
[project]
name = "mechanical-engineering-metrology-and-measurements"
version = "0.1.0"
description = "Jupyter book with Python snippets for the course Mechanical Engineering Metrology and Measurements"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ghp-import>=2.1.0",
    "jupyter-book>=1.0.4.post1",
    "jupytext>=1.18.1",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "pandoc>=2.4",
    "plotnine>=0.15.0",
    "pyyaml>=6.0.3",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
]
